requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
strawberry-graphql
python-dotenv
python-jose[cryptography]
pydantic
email-validator